+++ src/extension.ts
@@ -484,6 +484,16 @@ export function activate(context: vscode.ExtensionContext): void {
         })
     );
 
+    context.subscriptions.push(
+        vscode.commands.registerCommand('extension.vscode-janus-debug.showDocumentsVersion', () => {
+            if (loginData.documentsVersion && loginData.documentsVersion.length > 0) {
+                vscode.window.showInformationMessage(`DOCUMENTS Version ${loginData.documentsVersion}`);
+            } else {
+                vscode.window.showWarningMessage('Version only available after executing command on server');
+            }
+        })
+    );
+
     // connect the sever console manually
     context.subscriptions.push(
         vscode.commands.registerCommand('extension.vscode-janus.debug.connectServerConsole', (param) => {